+++ src/main/java/domain/GameStatistics.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participant.Participant;
 import java.util.EnumMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
