+++ src/main/java/domain/Participants.java
@@ -1,6 +1,9 @@
 package domain;
 
 import domain.card.Deck;
+import domain.participant.Hand;
+import domain.participant.Name;
+import domain.participant.Participant;
 import exception.ExceptionMessage;
 import java.util.ArrayList;
 import java.util.List;
