+++ src/main/java/domain/participant/Participant.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import static domain.GameResult.BLACKJACK_SCORE;
