+++ src/main/java/view/OutputView.java
@@ -3,9 +3,9 @@ package view;
 import domain.Dealer;
 import domain.GameResult;
 import domain.GameStatistics;
-import domain.Participant;
 import domain.Participants;
 import domain.card.Card;
+import domain.participant.Participant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
