+++ src/test/java/domain/HandTest.java
@@ -6,6 +6,7 @@ import domain.card.Card;
 import domain.card.Deck;
 import domain.card.Emblem;
 import domain.card.Grade;
+import domain.participant.Hand;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
